- avoid unnecessary API addition - resolve javadoc warning
