+++ org.eclipse.jdt.ui/ui/org/eclipse/jdt/internal/ui/wizards/buildpaths/ExternalAnnotationsAttachmentDialog.java
@@ -21,6 +21,7 @@ import org.eclipse.swt.widgets.Shell;
 import org.eclipse.core.runtime.IPath;
 
 import org.eclipse.jface.dialogs.StatusDialog;
+import org.eclipse.jface.window.Window;
 
 import org.eclipse.ui.PlatformUI;
 
@@ -31,11 +32,29 @@ import org.eclipse.jdt.internal.ui.wizards.IStatusChangeListener;
 import org.eclipse.jdt.internal.ui.wizards.NewWizardMessages;
 
 /**
- * A dialog to configure the external annotations attachment of a library (class folder, archives
- * and variable entries).
+ * A dialog to configure the external annotations attachment of any classpath entry.
  *
  */
 public class ExternalAnnotationsAttachmentDialog extends StatusDialog {
+
+	/**
+	 * Shows the UI for configuring an external annotations attachment. <code>null</code> is
+	 * returned when the user cancels the dialog. The dialog does not apply any changes.
+	 *
+	 * @param shell The parent shell for the dialog
+	 * @param initialEntry The entry to edit.
+	 * @param javaProject The enclosing java project, or {@code null} in which case a container
+	 * 	relative annotation path cannot be validated and has no content assist.
+	 * @return Returns the selected path, possibly different from the initialEntry,
+	 * or <code>null</code> if the dialog has been cancelled.
+	 */
+	public static IPath configureExternalAnnotationsAttachment(Shell shell, IPath initialEntry, IJavaProject javaProject) {
+		ExternalAnnotationsAttachmentDialog dialog= new ExternalAnnotationsAttachmentDialog(shell, initialEntry, javaProject);
+		if (dialog.open() == Window.OK) {
+			return dialog.getResult();
+		}
+		return null;
+	}
 
 	private ExternalAnnotationsAttachmentBlock fAnnotationsAttachmentBlock;
 
@@ -45,9 +64,10 @@ public class ExternalAnnotationsAttachmentDialog extends StatusDialog {
 	 *
 	 * @param parent Parent shell for the dialog
 	 * @param entry The entry to edit.
-	 * @param javaProject
+	 * @param javaProject The enclosing java project, or {@code null} in which case a container
+	 * 	relative annotation path cannot be validated and has no content assist.
 	 */
-	public ExternalAnnotationsAttachmentDialog(Shell parent, IPath entry, IJavaProject javaProject) {
+	protected ExternalAnnotationsAttachmentDialog(Shell parent, IPath entry, IJavaProject javaProject) {
 		super(parent);
 
 		IStatusChangeListener listener= this::updateStatus;
@@ -95,7 +115,7 @@ public class ExternalAnnotationsAttachmentDialog extends StatusDialog {
 	 *
 	 * @return the configured class path entry, or Path.EMPTY if no path was selected.
 	 */
-	public IPath getResult() {
+	protected IPath getResult() {
 		return fAnnotationsAttachmentBlock.getAnnotationsPath();
 	}
 }
